+++ doc/conf.py
@@ -27,10 +27,10 @@ author = "ixmp Developers"
 
 # -- General configuration -------------------------------------------------------------
 
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
+# Add any Sphinx extension module names here, as strings. They can be extensions coming
+# with Sphinx (named 'sphinx.ext.*') or your custom ones.
 extensions = [
+    "ixmp.utils.sphinx_linkcode_github",
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
     "sphinx.ext.coverage",
@@ -46,13 +46,13 @@ extensions = [
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
+# List of patterns, relative to source directory, that match files and directories to
+# ignore when looking for source files. This pattern also affects html_static_path and
+# html_extra_path.
 exclude_patterns = ["_build", "README.rst", "Thumbs.db", ".DS_Store"]
 
-# A string of reStructuredText that will be included at the beginning of every
-# source file that is read.
+# A string of reStructuredText that will be included at the beginning of every source
+# file that is read.
 version = ixmp.__version__
 rst_prolog = r"""
 .. |MESSAGEix| replace:: MESSAGE\ :emphasis:`ix`
@@ -76,24 +76,23 @@ html_css_files = ["custom.css"]
 
 html_favicon = "_static/favicon.svg"
 
-# The name of an image file (relative to this directory) to place at the top
-# of the sidebar.
+# The name of an image file (relative to this directory) to place at the top of the
+# sidebar.
 html_logo = "_static/combined-logo-white.png"
 
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
+# Add any paths that contain custom static files (such as style sheets) here, relative
+# to this directory. They are copied after the builtin static files, so a file named
+# "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
 
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
+# The theme to use for HTML and HTML Help pages.
 html_theme = "sphinx_rtd_theme"
 
 # -- Options for sphinx.ext.extlinks ---------------------------------------------------
 
 extlinks = {
-    "issue": ("https://github.com/iiasa/ixmp/issue/%s", "#"),
-    "pull": ("https://github.com/iiasa/ixmp/pull/%s", "PR #"),
+    "issue": ("https://github.com/iiasa/ixmp/issue/%s", "#%s"),
+    "pull": ("https://github.com/iiasa/ixmp/pull/%s", "PR #%s"),
 }
 
 # -- Options for sphinx.ext.intersphinx ------------------------------------------------
@@ -108,8 +107,13 @@ intersphinx_mapping = {
     "pint": ("https://pint.readthedocs.io/en/stable/", None),
     "python": ("https://docs.python.org/3/", None),
     "sparse": ("https://sparse.pydata.org/en/stable/", None),
+    "sphinx": ("https://www.sphinx-doc.org/en/master/", None),
     "xarray": ("https://xarray.pydata.org/en/stable/", None),
 }
+
+# -- Options for sphinx.ext.linkcode / ixmp.utils.sphinx_linkcode_github ---------------
+
+linkcode_github_repo_slug = "iiasa/ixmp"
 
 # -- Options for sphinx.ext.todo -------------------------------------------------------